+++ extras/test/cores/avr/Stream.h
@@ -25,11 +25,11 @@ struct Stream : Print {
   }
 
   String readString() {
-    size_t n = available();
-    std::string result;
-    result.resize(n);
-    n = readBytes(&result[0], n);
-    result.resize(n);
+    String result;
+    int c;
+    while ((c = read()) >= 0) {
+      result += static_cast<char>(c);
+    }
     return result;
   }
 };