+++ src/StreamUtils.h
@@ -14,6 +14,7 @@
 #include "StreamUtils/Streams/ReadBufferingStream.hpp"
 #include "StreamUtils/Streams/ReadLoggingStream.hpp"
 #include "StreamUtils/Streams/ReadThrottlingStream.hpp"
+#include "StreamUtils/Streams/StringStream.hpp"
 #include "StreamUtils/Streams/WriteBufferingStream.hpp"
 #include "StreamUtils/Streams/WriteLoggingStream.hpp"